Update to work on CIAO 4.9
The include file (array.hh and glue.hh) has been updated to match the
version from Sherpa in CIAO 4.9, which fixes several problems
when building the code with a modern C++ compiler.

To do this I had to special case the 1D model just for double arrays,
rather than let the array type be sent in to the template. There may
well be a better way to fix this issue.

The 2D model case has not been updated (i.e. it will need specializing
too).

It appears to work on Python 3.5 and 2.7.